Bug fixes & Updated Build
In level 2 a blockage has been moved to prevent an unintended clear route that only required 3 ninjas instead of 5.

Fixed an issue where upon completing a level, the "Soap Collected" text on the UI only showed the first digit of soap collected.

Fixed an issue where the "R" key was not properly respawning the ninjas in some levels.

Updated build