LLS Title, Bug Fix, Updated Build
Added a "Choose a level" title across the level selection screen

Fixed an issue where if you cleared a level and tried to go to the next, it would send you back to the previous level.

The secret reward for clearing all levels under the 100% category is now unlockable.

Added a check to see if the player 100%'s every level (not including the tutorial)

Updated Build.