Bug Fix, Sounds, Music, Cutscenes, Secret Ending
Fixed an issue where if a player completed a level and pressed the "R" key while the completion UI was active, the game would not function properly.

Fixed a text issue with the tutorial text.

Added background music to the level selection screen and all levels.

Added victory music for when a player completes a level.

Added a sound effect that plays when you collect soap.

Added intro cutscene

Fixed issue where some music started after an awkward delay.

Fixed an issue where the ending cutscene would not play.